Extract arrow decoding support to its own module.
