+++ pg_search/src/postgres/types.rs
@@ -37,6 +37,10 @@ use std::str::FromStr;
 use tantivy::schema::{IntoIpv6Addr, OwnedValue};
 use thiserror::Error;
 
+/// A row-oriented wrapper around Tantivy's OwnedValue.
+///
+/// When working with large batches of TantivyValues, consider using the `types_arrow` module
+/// instead.
 #[derive(Clone, Debug, Eq, PartialEq, PostgresType)]
 pub struct TantivyValue(pub tantivy::schema::OwnedValue);
